kit_trade_chatbot/requirements.txt
Flask
gunicorn
requests
line-bot-sdk
settrade
binance
